requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
efficientnet==1.1.1
gensim==3.8.3
googletrans==4.0.0-rc1
grad-cam == 1.3.7
h5py==3.1.0
imgaug==0.4.0
matplotlib==3.5.1
# nlg_eval==2.3
nltk==3.4.5
numpy==1.19.5
opencv-python-headless 
pandas == 1.4.2
plotly == 5.8.0
requests == 2.27.1
scikit_image == 0.19.2
scikit_learn == 1.0.2
seaborn == 0.11.2
streamlit == 1.8.1
streamlit-option-menu == 0.3.2
tensorflow==2.5.3
termcolor == 1.1.0
torch == 1.11.0
torchsummary == 1.5.1
torchvision == 0.12.0
torchxrayvision == 0.0.32
transformers == 2.5.1
tqdm==4.64.0
Pillow == 9.1.0
protobuf == 3.19.0

git+https://github.com/Maluuba/nlg-eval.git@master
